fix: extract sizing functions to separate file for Fast Refresh
Move estimateDecisionSize and estimateStartNodeWidth to saga-flow-sizing.ts
to satisfy the eslint rule that component files must only export components.
