+++ frontend/src/features/cookbook/components/saga-flow.tsx
@@ -13,6 +13,7 @@ import '@xyflow/react/dist/style.css'
 import Dagre from '@dagrejs/dagre'
 import type { SagaFlow } from '../lib/star-parser'
 import { parseTriggerService } from '../lib/star-parser'
+import { estimateDecisionSize, estimateStartNodeWidth } from './saga-flow-sizing'
 
 // Curated palette of visually distinct service colors using CSS custom properties
 const SERVICE_PALETTE = [
@@ -197,17 +198,6 @@ interface DecisionNodeData {
   [key: string]: unknown
 }
 
-/** Estimate diamond dimensions based on label length. The diamond clip-path
- *  only exposes ~50% of width/height for text, so we scale up for longer labels. */
-export function estimateDecisionSize(label: string): { width: number; height: number } {
-  const charWidth = 6 // approximate px per character at text-[10px]
-  const textWidth = label.length * charWidth
-  // Diamond usable area is ~50% of dimensions, add padding
-  const width = Math.max(120, Math.min(220, textWidth * 2 + 40))
-  const height = Math.max(80, Math.round(width * 0.7))
-  return { width, height }
-}
-
 const DecisionNode = memo(function DecisionNode({ data }: { data: DecisionNodeData }) {
   const dimmed = data.highlightedSaga && data.highlightedSaga !== data.sagaName
   const size = useMemo(() => estimateDecisionSize(data.label), [data.label])
@@ -292,14 +282,6 @@ const DEFAULT_NODE_DIMENSIONS: Record<string, { width: number; height: number }>
   sagaDecision: { width: 120, height: 80 },
   sagaExit: { width: 120, height: 36 },
   sagaEnd: { width: 140, height: 44 },
-}
-
-/** Estimate start node width based on the longer of name or trigger text. */
-export function estimateStartNodeWidth(label: string, trigger: string | null): number {
-  const charWidth = 6
-  const labelWidth = label.length * charWidth + 32 // px-4 padding
-  const triggerWidth = trigger ? trigger.length * charWidth + 16 : 0
-  return Math.max(160, Math.min(300, Math.max(labelWidth, triggerWidth)))
 }
 
 /** Get node dimensions, using dynamic sizing for decision and start nodes. */
